Deleting by index issue addressed
In this update I have made a fix for when the user removes elements from the list. Before, it used to remove the company even if the user searched for the 10'th company in the array, if it was the only search match and therefor was element 1 in the listBox, the program would remove the 1'st company from the list. This issue has now been resolved. It has only been implemented in one place, however in the next update I will be implemented across the board.

Also in the next update I will be updating my design to become more visually appealing and also remove blank space.