Implementation of item target algorithm
Now, instead of removing and editing by index, the program checks for a match of unique object summary. This allows the program to run better and for the user to actually target the correct object after a search.

Also I have added using statements where best suited. 

In the next update I will be working on the new added buttons along the bottom of the app.