Changes to auto save feature and methods
Method visibility has been updated (changed to private where possible).

Notification method has been changed to allow support for the new save feature. 

A new class has been added (empty in this patch) to allow for file services to be in their own class.

In the next update, all of the address will be adjusted to fit the container better.